fix: add defer to content.js, unify ContentFile type, remove dead code, fix onAfterRender empty string
